Add Authentication in Admin Pannel

+++ backend/controllers/foodController.js
@@ -1,4 +1,5 @@
 import foodModel from "../models/foodModel.js";
+import userModel from "../models/userModel.js";
 import fs from "fs";
 
 // add food items
@@ -13,8 +14,13 @@ const addFood = async (req, res) => {
     image: image_filename,
   });
   try {
-    await food.save();
-    res.json({ success: true, message: "Food Added" });
+    let userData = await userModel.findById(req.body.userId);
+    if (userData && userData.role === "admin") {
+      await food.save();
+      res.json({ success: true, message: "Food Added" });
+    } else {
+      res.json({ success: false, message: "You are not admin" });
+    }
   } catch (error) {
     console.log(error);
     res.json({ success: false, message: "Error" });
@@ -35,10 +41,15 @@ const listFood = async (req, res) => {
 // remove food item
 const removeFood = async (req, res) => {
   try {
-    const food = await foodModel.findById(req.body.id);
-    fs.unlink(`uploads/${food.image}`, () => {});
-    await foodModel.findByIdAndDelete(req.body.id);
-    res.json({ success: true, message: "Food Removed" });
+    let userData = await userModel.findById(req.body.userId);
+    if (userData && userData.role === "admin") {
+      const food = await foodModel.findById(req.body.id);
+      fs.unlink(`uploads/${food.image}`, () => {});
+      await foodModel.findByIdAndDelete(req.body.id);
+      res.json({ success: true, message: "Food Removed" });
+    } else {
+      res.json({ success: false, message: "You are not admin" });
+    }
   } catch (error) {
     console.log(error);
     res.json({ success: false, message: "Error" });
